создание товаров и ротация токенов а также карточка товаров детальная

+++ api/endpoints/auth.py
@@ -6,34 +6,46 @@ from adapters.web import GetUserDep
 from services.auth import create_tokens_from_login, create_tokens_from_refresh, user_register, delete_redis_keys
 from adapters.web import AuthCookies
 from infra.security import verify
+from infra.auth import get_data_from_token
 
-from .schemas import UserLogin, UserRegister
+from api.schemas import UserLogin, UserRegister
+import logging
 
 router = APIRouter()
+log = logging.getLogger(__name__)
 
-@router.get("/logout")
+@router.post("/logout")
 async def logout(request: Request, redis: RedisDep):
     cookie_manager = AuthCookies()
     response = JSONResponse(content="success")
     refresh_token = cookie_manager.get_refresh_token(request=request)
-    jti, family_id = refresh_token["jti"], refresh_token["family_id"]
+    if refresh_token:
+        refresh_token_data = get_data_from_token(refresh_token)
+        jti, family_id = refresh_token_data["jti"], refresh_token_data["family_id"]
+        await delete_redis_keys(redis=redis, jti=jti, family_id=family_id)
     cookie_manager.clear_tokens(response=response)
-    await delete_redis_keys(redis=redis, jti=jti, family_id=family_id)
     return response
 
 
 @router.get("/me")
-async def get_user_profile(user: GetUserDep):
+async def get_user_profile(request: Request, user: GetUserDep):
+    cookie_manager = AuthCookies()
+    access_token = cookie_manager.get_access_token(request=request)
+    #log.debug("access_token: %s", access_token)
     return {"user": user}
+
 
 @router.post("/refresh")
 async def refresh_tokens(request: Request, redis: RedisDep, uow: UowDep):
     cookie_manager = AuthCookies()
     refresh_token = cookie_manager.get_refresh_token(request=request)
+    # access_token = cookie_manager.get_access_token(request=request)
+    # log.debug("access_token: %s", access_token)
+    # log.debug("refresh_token: %s", refresh_token)
     tokens = await create_tokens_from_refresh(redis=redis, uow=uow, refresh_token=refresh_token)
     cookie_manager = AuthCookies()
     response = JSONResponse(content="success")
-    cookie_manager.set_tokens(tokens=tokens, response=response)
+    cookie_manager.set_tokens(**tokens, response=response)
     return response
 
 
@@ -42,7 +54,7 @@ async def user_login(user: UserLogin, uow: UowDep, redis: RedisDep):
     tokens, user_data = await create_tokens_from_login(uow, username=user.username, password=user.password, verify=verify, redis=redis)
     cookie_manager = AuthCookies()
     response = JSONResponse(content={"user": user_data})
-    cookie_manager.set_tokens(tokens=tokens, response=response)
+    cookie_manager.set_tokens(**tokens, response=response)
     return response
 
 
@@ -51,5 +63,5 @@ async def register(user: UserRegister, uow: UowDep, redis: RedisDep):
     tokens, user_data = await user_register(redis=redis, uow=uow, **user.model_dump())
     cookie_manager = AuthCookies()
     response = JSONResponse(content={"user": user_data})
-    cookie_manager.set_tokens(tokens=tokens, response=response)
+    cookie_manager.set_tokens(**tokens, response=response)
     return response
